requirements.txt
black==22.10.0
click==8.1.3
Flask==2.2.2
Flask-API==3.0.post1
Flask-Cors==3.0.10
gunicorn==20.1.0
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
psycopg2-binary==2.9.4
PyJWT==2.6.0
python-dotenv==0.21.0
six==1.16.0
tomli==2.0.1
typing_extensions==4.4.0
Werkzeug==2.2.2
zipp==3.9.0